Moved TAS and CGAC models out to a new domainModels file
