+++ dataactcore/models/validationModels.py
@@ -2,10 +2,7 @@
 
 from sqlalchemy import Column, Integer, Text, ForeignKey, Boolean
 from sqlalchemy.orm import relationship
-from sqlalchemy.ext.declarative import declarative_base
-from dataactcore.utils.timeStampMixin import TimeStampBase
-
-Base = declarative_base(cls=TimeStampBase)
+from dataactcore.models.domainModels import Base
 
 class FileType(Base):
     __tablename__ = "file_type"
@@ -74,19 +71,4 @@ class RuleTiming(Base):
 
     TIMING_DICT = None
 
-class TASLookup(Base) :
-    __tablename__ = "tas_lookup"
-    tas_id = Column(Integer, primary_key=True)
-    allocation_transfer_agency = Column(Text, nullable=True, index=True)
-    agency_identifier = Column(Text, nullable=True, index=True)
-    beginning_period_of_availability = Column(Text, nullable=True, index=True)
-    ending_period_of_availability = Column(Text, nullable=True, index=True)
-    availability_type_code = Column(Text, nullable=True, index=True)
-    main_account_code = Column(Text, nullable=True, index=True)
-    sub_account_code = Column(Text, nullable=True, index=True)
 
-class CGAC(Base):
-    __tablename__ = "cgac"
-    cgac_id = Column(Integer, primary_key=True)
-    cgac_code = Column(Text, nullable=False,index=True,unique=True)
-    agency_name = Column(Text)
